refactor: RaceCarGame 클래스 raceCount 변수 추가
