+++ src/main/java/racingcar/service/GameLauncher.java
@@ -9,19 +9,27 @@ import racingcar.domain.RacingCar;
 public class GameLauncher {
 
     private final List<Car> racingCarList;
+    private final int endRaceCount;
+    private int currentCount = 0;
 
-    public GameLauncher(List<String> nameList) {
+    public GameLauncher(List<String> nameList, int endRaceCount) {
         racingCarList = new ArrayList<>();
-
         for (String name : nameList) {
             racingCarList.add(new RacingCar(name));
         }
+
+        this.endRaceCount = endRaceCount;
     }
 
     public void moveForwardAll() {
+        currentCount ++;
         for (Car car : racingCarList) {
             car.moveForward();
         }
+    }
+
+    public boolean isEnd() {
+        return endRaceCount <= currentCount;
     }
 
     public List<Car> getRacingCarList() {
