+++ src/test/java/calculator/view/FormulaInputTest.java
@@ -1,15 +1,9 @@
-package calculator.ui;
+package calculator.view;
 
 import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.PrintStream;
 import java.util.Arrays;
 import java.util.List;
-import org.assertj.core.api.Assert;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
